+++ src/apps/base.js
@@ -2,7 +2,7 @@ const express = require('express');
 const cors = require('cors');
 
 const swaggerUi = require('swagger-ui-express');
-const swaggerDocument = require('./swagger.json');
+const swaggerDocument = require('../swagger.json');
 
 
 // eslint-disable-next-line new-cap
