+++ src/apps/salary.js
